Refactor Breeze to group related methods and classes together
This change refactors Breeze classes to more consistent approach.

* The "commands" package only contains commands
* All Parameters (BuildCi, BuildProd, BuildDoc, Shell) are now
  in "params" package
* Required/Optional Build args are now members of the
  BuildCiParams, BuildProdParams which makes the params
  much more self-contained..
* All utils are in "utils" package

This helps with avoiding circular imports (all utios are now
standalone and do not use any of the commands.
