+++ dev/breeze/src/airflow_breeze/commands/developer_commands.py
@@ -16,14 +16,25 @@
 # under the License.
 
 import os
+import subprocess
 import sys
-from dataclasses import dataclass
-from typing import List, Optional, Tuple
+from typing import Optional, Tuple, Union
 
 import rich_click as click
 
-from airflow_breeze.build_image.ci.build_ci_params import BuildCiParams
-from airflow_breeze.commands.common_options import (
+from airflow_breeze.commands.ci_image_commands import rebuild_ci_image_if_needed
+from airflow_breeze.commands.main_command import main
+from airflow_breeze.global_constants import (
+    DEFAULT_PYTHON_MAJOR_MINOR_VERSION,
+    MOUNT_SELECTED,
+    get_available_packages,
+)
+from airflow_breeze.params.build_ci_params import BuildCiParams
+from airflow_breeze.params.doc_build_params import DocBuildParams
+from airflow_breeze.params.shell_params import ShellParams
+from airflow_breeze.pre_commit_ids import PRE_COMMIT_LIST
+from airflow_breeze.utils.cache import read_from_cache_file
+from airflow_breeze.utils.common_options import (
     option_airflow_constraints_reference,
     option_airflow_extras,
     option_answer,
@@ -47,24 +58,19 @@ from airflow_breeze.commands.common_options import (
     option_use_packages_from_dist,
     option_verbose,
 )
-from airflow_breeze.commands.custom_param_types import BetterChoice
-from airflow_breeze.commands.main import main
-from airflow_breeze.global_constants import (
-    DEFAULT_PYTHON_MAJOR_MINOR_VERSION,
-    MOUNT_SELECTED,
-    get_available_packages,
-)
-from airflow_breeze.pre_commit_ids import PRE_COMMIT_LIST
-from airflow_breeze.shell.enter_shell import enter_shell, find_airflow_container
-from airflow_breeze.shell.shell_params import ShellParams
 from airflow_breeze.utils.console import get_console
+from airflow_breeze.utils.custom_param_types import BetterChoice
 from airflow_breeze.utils.docker_command_utils import (
+    check_docker_compose_version,
+    check_docker_is_running,
+    check_docker_resources,
+    check_docker_version,
     get_env_variables_for_docker_commands,
     get_extra_docker_flags,
 )
 from airflow_breeze.utils.path_utils import AIRFLOW_SOURCES_ROOT
-from airflow_breeze.utils.rebuild_image_if_needed import rebuild_ci_image_if_needed
-from airflow_breeze.utils.run_utils import assert_pre_commit_installed, run_command
+from airflow_breeze.utils.run_utils import assert_pre_commit_installed, filter_out_none, run_command
+from airflow_breeze.utils.visuals import ASCIIART, ASCIIART_STYLE, CHEATSHEET, CHEATSHEET_STYLE
 
 DEVELOPER_COMMANDS = {
     "name": "Developer tools",
@@ -207,6 +213,8 @@ DEVELOPER_PARAMETERS = {
 # Is used for a shorthand of shell and except the extra
 # Args it should have the same parameters.
 # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+
 @main.command()
 @option_verbose
 @option_dry_run
@@ -516,30 +524,6 @@ def stop(verbose: bool, dry_run: bool, preserve_volumes: bool):
     run_command(command_to_execute, verbose=verbose, dry_run=dry_run, env=env_variables)
 
 
-@dataclass
-class DocBuildParams:
-    package_filter: Tuple[str]
-    docs_only: bool
-    spellcheck_only: bool
-    for_production: bool
-    skip_environment_initialization: bool = False
-    github_actions = os.environ.get('GITHUB_ACTIONS', "false")
-
-    @property
-    def args_doc_builder(self) -> List[str]:
-        doc_args = []
-        if self.docs_only:
-            doc_args.append("--docs-only")
-        if self.spellcheck_only:
-            doc_args.append("--spellcheck-only")
-        if self.for_production:
-            doc_args.append("--for-production")
-        if self.package_filter and len(self.package_filter) > 0:
-            for single_filter in self.package_filter:
-                doc_args.extend(["--package-filter", single_filter])
-        return doc_args
-
-
 @main.command(name='exec', help='Joins the interactive shell of running airflow container')
 @option_verbose
 @option_dry_run
@@ -567,3 +551,101 @@ def exec(verbose: bool, dry_run: bool, exec_args: Tuple):
         if not process:
             sys.exit(1)
         sys.exit(process.returncode)
+
+
+def enter_shell(**kwargs) -> Union[subprocess.CompletedProcess, subprocess.CalledProcessError]:
+    """
+    Executes entering shell using the parameters passed as kwargs:
+
+    * checks if docker version is good
+    * checks if docker-compose version is good
+    * updates kwargs with cached parameters
+    * displays ASCIIART and CHEATSHEET unless disabled
+    * build ShellParams from the updated kwargs
+    * executes the command to drop the user to Breeze shell
+
+    """
+    verbose = kwargs['verbose']
+    dry_run = kwargs['dry_run']
+    check_docker_is_running(verbose)
+    check_docker_version(verbose)
+    check_docker_compose_version(verbose)
+    if read_from_cache_file('suppress_asciiart') is None:
+        get_console().print(ASCIIART, style=ASCIIART_STYLE)
+    if read_from_cache_file('suppress_cheatsheet') is None:
+        get_console().print(CHEATSHEET, style=CHEATSHEET_STYLE)
+    enter_shell_params = ShellParams(**filter_out_none(**kwargs))
+    return run_shell_with_build_image_checks(verbose, dry_run, enter_shell_params)
+
+
+def run_shell_with_build_image_checks(
+    verbose: bool, dry_run: bool, shell_params: ShellParams
+) -> Union[subprocess.CompletedProcess, subprocess.CalledProcessError]:
+    """
+    Executes a shell command built from params passed, checking if build is not needed.
+    * checks if there are enough resources to run shell
+    * checks if image was built at least once (if not - forces the build)
+    * if not forces, checks if build is needed and asks the user if so
+    * builds the image if needed
+    * prints information about the build
+    * constructs docker compose command to enter shell
+    * executes it
+
+    :param verbose: print commands when running
+    :param dry_run: do not execute "write" commands - just print what would happen
+    :param shell_params: parameters of the execution
+    """
+    rebuild_ci_image_if_needed(build_params=shell_params, dry_run=dry_run, verbose=verbose)
+    shell_params.print_badge_info()
+    cmd = ['docker-compose', 'run', '--service-ports', "-e", "BREEZE", '--rm', 'airflow']
+    cmd_added = shell_params.command_passed
+    env_variables = get_env_variables_for_docker_commands(shell_params)
+    if cmd_added is not None:
+        cmd.extend(['-c', cmd_added])
+
+    command_result = run_command(
+        cmd, verbose=verbose, dry_run=dry_run, env=env_variables, text=True, check=False
+    )
+    if command_result.returncode == 0:
+        return command_result
+    else:
+        get_console().print(f"[red]Error {command_result.returncode} returned[/]")
+        if verbose:
+            get_console().print(command_result.stderr)
+        return command_result
+
+
+def stop_exec_on_error(returncode: int):
+    get_console().print('\n[error]ERROR in finding the airflow docker-compose process id[/]\n')
+    sys.exit(returncode)
+
+
+def find_airflow_container(verbose, dry_run) -> Optional[str]:
+    exec_shell_params = ShellParams(verbose=verbose, dry_run=dry_run)
+    check_docker_resources(exec_shell_params.airflow_image_name, verbose=verbose, dry_run=dry_run)
+    exec_shell_params.print_badge_info()
+    env_variables = get_env_variables_for_docker_commands(exec_shell_params)
+    cmd = ['docker-compose', 'ps', '--all', '--filter', 'status=running', 'airflow']
+    docker_compose_ps_command = run_command(
+        cmd, verbose=verbose, dry_run=dry_run, text=True, capture_output=True, env=env_variables, check=False
+    )
+    if dry_run:
+        return "CONTAINER_ID"
+    if docker_compose_ps_command.returncode != 0:
+        if verbose:
+            get_console().print(docker_compose_ps_command.stdout)
+            get_console().print(docker_compose_ps_command.stderr)
+        stop_exec_on_error(docker_compose_ps_command.returncode)
+        return None
+
+    output = docker_compose_ps_command.stdout
+    container_info = output.strip().split('\n')
+    if container_info:
+        container_running = container_info[-1].split(' ')[0]
+        if container_running.startswith('-'):
+            # On docker-compose v1 we get '--------' as output here
+            stop_exec_on_error(docker_compose_ps_command.returncode)
+        return container_running
+    else:
+        stop_exec_on_error(1)
+        return None
