+++ dev/breeze/src/airflow_breeze/commands/production_image_commands.py
@@ -14,15 +14,17 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
+import contextlib
+import os
 import sys
 from typing import Optional, Tuple
 
 import click
 
-from airflow_breeze.build_image.prod.build_prod_image import build_production_image
-from airflow_breeze.build_image.prod.build_prod_params import BuildProdParams
-from airflow_breeze.commands.common_options import (
+from airflow_breeze.commands.main_command import main
+from airflow_breeze.global_constants import ALLOWED_INSTALLATION_METHODS, DEFAULT_EXTRAS
+from airflow_breeze.params.build_prod_params import BuildProdParams
+from airflow_breeze.utils.common_options import (
     option_additional_dev_apt_command,
     option_additional_dev_apt_deps,
     option_additional_dev_apt_env,
@@ -61,14 +63,19 @@ from airflow_breeze.commands.common_options import (
     option_verify_image,
     option_wait_for_image,
 )
-from airflow_breeze.commands.custom_param_types import BetterChoice
-from airflow_breeze.commands.main import main
-from airflow_breeze.global_constants import ALLOWED_INSTALLATION_METHODS, DEFAULT_EXTRAS
 from airflow_breeze.utils.console import get_console
+from airflow_breeze.utils.custom_param_types import BetterChoice
+from airflow_breeze.utils.docker_command_utils import (
+    construct_docker_build_command,
+    construct_empty_docker_build_command,
+    tag_and_push_image,
+)
+from airflow_breeze.utils.path_utils import AIRFLOW_SOURCES_ROOT, DOCKER_CONTEXT_DIR
 from airflow_breeze.utils.pulll_image import run_pull_image, run_pull_in_parallel
 from airflow_breeze.utils.python_versions import get_python_version_list
+from airflow_breeze.utils.registry import login_to_docker_registry
 from airflow_breeze.utils.run_tests import verify_an_image
-from airflow_breeze.utils.run_utils import filter_out_none
+from airflow_breeze.utils.run_utils import filter_out_none, fix_group_permissions, run_command
 
 PRODUCTION_IMAGE_TOOLS_COMMANDS = {
     "name": "Production Image tools",
@@ -387,3 +394,124 @@ def verify_prod_image(
         extra_pytest_args=extra_pytest_args,
     )
     sys.exit(return_code)
+
+
+def clean_docker_context_files(verbose: bool, dry_run: bool):
+    """
+    Cleans up docker context files folder - leaving only .README.md there.
+    """
+    if verbose or dry_run:
+        get_console().print("[info]Cleaning docker-context-files[/]")
+    if dry_run:
+        return
+    with contextlib.suppress(FileNotFoundError):
+        context_files_to_delete = DOCKER_CONTEXT_DIR.glob('**/*')
+        for file_to_delete in context_files_to_delete:
+            if file_to_delete.name != '.README.md':
+                file_to_delete.unlink()
+
+
+def check_docker_context_files(install_packages_from_context: bool):
+    """
+    Sanity check - if we want to install from docker-context-files we expect some packages there but if
+    we don't - we don't expect them, and they might invalidate Docker cache.
+
+    This method exits with an error if what we see is unexpected for given operation.
+
+    :param install_packages_from_context: whether we want to install from docker-context-files
+    """
+    context_file = DOCKER_CONTEXT_DIR.glob('**/*')
+    number_of_context_files = len(
+        [context for context in context_file if context.is_file() and context.name != '.README.md']
+    )
+    if number_of_context_files == 0:
+        if install_packages_from_context:
+            get_console().print('[warning]\nERROR! You want to install packages from docker-context-files')
+            get_console().print('[warning]\n but there are no packages to install in this folder.')
+            sys.exit(1)
+    else:
+        if not install_packages_from_context:
+            get_console().print(
+                '[warning]\n ERROR! There are some extra files in docker-context-files except README.md'
+            )
+            get_console().print('[warning]\nAnd you did not choose --install-packages-from-context flag')
+            get_console().print(
+                '[warning]\nThis might result in unnecessary cache invalidation and long build times'
+            )
+            get_console().print(
+                '[warning]\nExiting now \
+                    - please restart the command with --cleanup-context switch'
+            )
+            sys.exit(1)
+
+
+def build_production_image(
+    verbose: bool, dry_run: bool, prod_image_params: BuildProdParams
+) -> Tuple[int, str]:
+    """
+    Builds PROD image:
+
+      * fixes group permissions for files (to improve caching when umask is 002)
+      * converts all the parameters received via kwargs into BuildProdParams (including cache)
+      * prints info about the image to build
+      * removes docker-context-files if requested
+      * performs sanity check if the files are present in docker-context-files if expected
+      * logs int to docker registry on CI if build cache is being executed
+      * removes "tag" for previously build image so that inline cache uses only remote image
+      * constructs docker-compose command to run based on parameters passed
+      * run the build command
+      * update cached information that the build completed and saves checksums of all files
+        for quick future check if the build is needed
+
+    :param verbose: print commands when running
+    :param dry_run: do not execute "write" commands - just print what would happen
+    :param prod_image_params: PROD image parameters
+    """
+    fix_group_permissions(verbose=verbose)
+    if verbose or dry_run:
+        get_console().print(
+            f"\n[info]Building PROD image of airflow from {AIRFLOW_SOURCES_ROOT} "
+            f"python version: {prod_image_params.python}[/]\n"
+        )
+    if prod_image_params.cleanup_context:
+        clean_docker_context_files(verbose=verbose, dry_run=dry_run)
+    check_docker_context_files(prod_image_params.install_packages_from_context)
+    if prod_image_params.prepare_buildx_cache:
+        login_to_docker_registry(prod_image_params, dry_run=dry_run)
+    run_command(
+        ["docker", "rmi", "--no-prune", "--force", prod_image_params.airflow_image_name],
+        verbose=verbose,
+        dry_run=dry_run,
+        cwd=AIRFLOW_SOURCES_ROOT,
+        text=True,
+        check=False,
+    )
+    get_console().print(f"\n[info]Building PROD Image for Python {prod_image_params.python}\n")
+    if prod_image_params.empty_image:
+        env = os.environ.copy()
+        env['DOCKER_BUILDKIT'] = "1"
+        get_console().print(f"\n[info]Building empty PROD Image for Python {prod_image_params.python}\n")
+        cmd = construct_empty_docker_build_command(image_params=prod_image_params)
+        build_command_result = run_command(
+            cmd,
+            input="FROM scratch\n",
+            verbose=verbose,
+            dry_run=dry_run,
+            cwd=AIRFLOW_SOURCES_ROOT,
+            check=False,
+            text=True,
+            env=env,
+        )
+    else:
+        cmd = construct_docker_build_command(
+            image_params=prod_image_params,
+            verbose=verbose,
+            production_image=True,
+        )
+        build_command_result = run_command(
+            cmd, verbose=verbose, dry_run=dry_run, cwd=AIRFLOW_SOURCES_ROOT, check=False, text=True
+        )
+    if build_command_result.returncode == 0:
+        if prod_image_params.push_image:
+            return tag_and_push_image(image_params=prod_image_params, dry_run=dry_run, verbose=verbose)
+    return build_command_result.returncode, f"Image build: {prod_image_params.python}"
