+++ dev/breeze/src/airflow_breeze/params/shell_params.py
@@ -14,7 +14,7 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Breeze shell paameters."""
+
 import os
 from dataclasses import dataclass
 from pathlib import Path
